Restructure code into packages
